move stuff around to avoid extra deps and simplify providers
