+++ packages/provider-s3/package.json
@@ -22,11 +22,7 @@
     "@aws-sdk/credential-providers": "^3.830.0",
     "@aws-sdk/s3-request-presigner": "^3.830.0",
     "@rnef/tools": "^0.8.7",
-    "adm-zip": "^0.5.16",
     "tslib": "^2.3.0"
-  },
-  "devDependencies": {
-    "@types/adm-zip": "^0.5.7"
   },
   "publishConfig": {
     "access": "public"
